ms_back_end/requirements.txt
Django==3.2.13
django-filter==21.1
django-guardian==2.4.0
djangorestframework==3.13.1
pillow==9.4.0
django-cors-headers==3.10.1
requests==2.27.1
